pyproject.toml
[project]
name = "advisor-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "e2b-code-interpreter>=2.3.0",
    "fastapi>=0.121.0",
    "firecrawl-py>=4.6.0",
    "google-generativeai>=0.8.5",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
